Modify lines
Show exact line number on the line the cursor is on and relative line
number otherwise - especially when files are larger, jumping to
different lines on the screen is easier.

Set textwidth to 99 - apparently having one less than what you want is
good and I like 100 for lines of code. 80 is default terminal size but
too short in my opinion. Also I have a 15" scren.